Peg clock value to whisper timestamp
This commit pegs the clock value to +/- 120 seconds from the whisper
timestamp.
In this way the scope to re-order messages in the past/future is limited
to 120s.

Membership updates are a special case, as those can be relayed without
the initial whisper message, so we can only check that they don't fall
too far in the future, but we can't check them against re-ordering of
past events. The scope in this case is limited as the protocol has been
built so that it's not a problem (current status-react behavior is the
same).